+++ docs/changelog.rst
@@ -14,7 +14,7 @@ E7EPD Database Specification
 * v0.2:
     * Added Inductor table
     * Added Diode table
-* v0.3 (In Progress, unreleased):
+* v0.3:
     * Added a ``storage`` key to all parts
     * Merged ``user_comments`` and ``part_comments`` to just one ``comments`` column
     * Removed ``power`` for the capacitor table
@@ -45,7 +45,7 @@ Database Python DB Wrapper
     * Added ```wipe_database``` function
     * Added a way for the backend to store configurations about itself
     * Added a key to check the database specification the database is under and the ``E7EPD`` class
-* v0.3 (In Progress, unreleased):
+* v0.3:
     * Switched to sqlalchemy for handing SQL
     * As the wrapper input includes a sqlalchemy engine, any sql type that sqlalchemy supports should be supported
     * Updated spec for Database Rev 0.3
@@ -67,8 +67,9 @@ CLI
     * Added option to remove and append stock to a part
     * Moved around options so that there is an "initial screen" before choosing components
     * Added a check for the database revision on startup
-* v0.3 (In Progress, unreleased)
+* v0.3:
     * Updated for the new Wrapper 0.3 database argument
+    * Changed options so it's easier to add a new part
     * Allowing option for a mySQL database
     * Allowing option to add multiple databases
     * Allowing option to select which database to connect to